Inject auth context into tanstack router (#23)
