+++ packages/app/src/components/signed-out/login-form.tsx
@@ -2,11 +2,12 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { useAuth } from '@/context/AuthContext';
+import { useAuth } from '@/context/auth-context';
 import { cn } from '@/lib/utils';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Link, useNavigate } from '@tanstack/react-router';
 import { useForm } from 'react-hook-form';
+import { toast } from 'sonner';
 import { z } from 'zod';
 
 const loginSchema = z.object({
@@ -30,10 +31,15 @@ function LoginForm({ className, ...props }: Readonly<React.ComponentPropsWithout
 
   const onSubmit = async (credentials: LoginFormValues) => {
     try {
-      await signIn(credentials);
-      navigate({ to: '/daily-log' });
+      const result = await signIn(credentials);
+      if (result) {
+        navigate({ to: '/daily-log' });
+      } else {
+        toast.error('Login failed');
+      }
     } catch (error) {
       console.error('Login failed:', error);
+      toast.error('Login failed with error');
     }
   };
